service/requirements.txt
discord
python-dateutil
python-dotenv
pytz
requests
